Fixed bugs with POST request for new order. Added checking for response dictionary in OrdersDataParser. Because I use the parser for a new order response. Added some changes into OrdersViewController.

+++ LV-116.iOS.SS.Restaurant/LV-116.iOS.SS.Restaurant/Reparser/ParserToJSON.m
@@ -57,8 +57,8 @@ static NSString *const kKeyUserId       =  @"UserId";
 +(NSData*)createJSONDataForNewOrderWithTableId:(int)tableId
 {
     NSNumber *tableIdNumber = [NSNumber numberWithInt: tableId];
-      long int timeInMiliseconds = [[NSDate date] timeIntervalSince1970] * 1000;
-    NSString *timestamp = [[NSString alloc] initWithFormat:@"/Date(%ld+0200)/", timeInMiliseconds];
+      long long int timeInMiliseconds = [[NSDate date] timeIntervalSince1970] * 1000;
+    NSString *timestamp = [[NSString alloc] initWithFormat:@"/Date(%lld+0200)/", timeInMiliseconds];
 //    NSNumber *userIdNumber = [NSNumber numberWithInt: ([UserRole getInstance]).userId ];
     NSNumber *userIdNumber = @1049;
